// frontend/src/API.ts
export enum Months {
  JAN = "JAN",
  FEB = "FEB",
  MAR = "MAR",
  APR = "APR",
  MAY = "MAY",
  JUN = "JUN",
  JUL = "JUL",
  AUG = "AUG",
  SEP = "SEP",
  OCT = "OCT",
  NOV = "NOV",
  DEC = "DEC",
}
// ...
export type CreativeRequestEarnings = {
  __typename: "CreativeRequestEarnings",
  creativeRequestEarningId: string,
  creativeRequestId: string,
  creatorId: string,
  amount: number,
  creativeUniqueId: string,
  currentEarnings: number,
  month: Months,
  toDate: string,
  fromDate: string,
  createdAt: string,
  updatedAt: string,
};
// ...
export type AddCreativeEarningMutationVariables = {
  userProfileId: string,
  creativeRequestId: string,
  amount: number,
  month: Months,
  toDate: string,
  fromDate: string,
};